refactor(research): isolate mission action lock
